pyproject.toml
[tool.poetry]
name = "django-example-project"
version = "0.1.0"
description = ""
authors = ["Efficax Team <efficax.team@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-dotenv = "^1.4.2"
djangorestframework = "^3.14.0"
uuid = "^1.30"
drf-yasg = "^1.21.5"
daphne = "^4.0.0"
django-cors-headers = "^3.14.0"
python-jose = "^3.3.0"
django-filter = "^23.2"
channels = "^4.0.0"
channels-redis = "^4.1.0"
django-celery-beat = "^2.5.0"
python-crontab = "^2.7.1"
django-timezone-field = "^5.0"
python-dateutil = "^2.8.2"
celery = "^5.2.7"
django-celery-results = "^2.5.0"
redis = "^4.5.4"
twisted = {extras = ["http2", "tls"], version = "^22.10.0"}
loguru = "^0.7.0"
black = "^23.3.0"

[tool.poetry.group.windows]
optional = true

[tool.poetry.group.windows.dependencies]
psycopg2 = "^2.9.6"

[tool.poetry.group.linux]
optional = true

[tool.poetry.group.linux.dependencies]
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
